+++ src/Yayaml/Yaml12Schema.cs
@@ -1,7 +1,6 @@
 using System;
 using System.Globalization;
 using System.Linq;
-using System.Management.Automation;
 using System.Numerics;
 using System.Text.RegularExpressions;